# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Calendros API Documentation",
        "description": "API documentation for Calendros - A comprehensive calendar management system",
        "contact": {
            "email": "support@calendros.example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://http://calendros.test/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://prm.positive.io.vn/api",
            "description": "Production Server"
        },
        {
            "url": "https://calendros.montserrat.id.vn/api",
            "description": "Test Server"
        }
    ],
    "paths": {
...
        "/calendars": {
            "get": {
                "tags": [
                    "Calendars"
                ],
                "summary": "List user's calendars",
                "description": "Get all calendars owned by the authenticated user, ordered by default status and name",
                "operationId": "e9180bba7d2715f47fcc2c79913890d8",
                "responses": {
                    "200": {
                        "description": "List of calendars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Personal Calendar"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "My personal events"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "#4F46E5"
                                                    },
                                                    "timezone": {
                                                        "type": "string",
                                                        "example": "UTC"
                                                    },
                                                    "is_default": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Create a new calendar",
                "description": "Create a new calendar for the authenticated user",
                "operationId": "be760f5383cac94d3df8d0b9d0a34c81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Work Calendar"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Work-related events"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#10B981"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "America/New_York"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Calendar created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Work Calendar"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Work-related events"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#10B981"
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "America/New_York"
                                                },
                                                "is_default": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/calendars/{id}": {
            "get": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Get calendar details",
                "description": "Retrieve details of a specific calendar (must be owned by authenticated user)",
                "operationId": "255ba7f7f9fea6ad7ef00d915aa3e2b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Calendar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendar details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Personal Calendar"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "My personal events"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#4F46E5"
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "UTC"
                                                },
                                                "is_default": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not your calendar"
                    },
                    "404": {
                        "description": "Calendar not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Update calendar",
                "description": "Update an existing calendar (must be owned by authenticated user)",
                "operationId": "ef8006d6016af06f555fdf4bc1ff1682",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Calendar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Calendar"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#EF4444"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Europe/London"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calendar updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Updated Calendar"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Calendar not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Delete calendar",
                "description": "Delete a calendar (must be owned by authenticated user). All associated events will also be deleted.",
                "operationId": "8a8814c27a5c100cc5d870a25a83d623",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Calendar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Calendar deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Calendar not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}